don't try to create link if it already exists

+++ clang-build/build.sh
@@ -4,5 +4,9 @@ echo running build.sh...
 set -e
 
 cd /mnt/wasi-sdk
-ln -s /mnt/llvm-project src/llvm-project
+
+if ! [ -L src/llvm-project ]; then
+    ln -s /mnt/llvm-project src/llvm-project
+fi
+
 make
